+++ apps/api/src/app/admin/admin.module.ts
@@ -1,4 +1,5 @@
 import { SubscriptionModule } from '@ghostfolio/api/app/subscription/subscription.module';
+import { TransformDataSourceInRequestModule } from '@ghostfolio/api/interceptors/transform-data-source-in-request/transform-data-source-in-request.module';
 import { ApiModule } from '@ghostfolio/api/services/api/api.module';
 import { ConfigurationModule } from '@ghostfolio/api/services/configuration/configuration.module';
 import { DataGatheringModule } from '@ghostfolio/api/services/data-gathering/data-gathering.module';
@@ -29,7 +30,8 @@ import { QueueModule } from './queue/queue.module';
     QueueModule,
     SubscriptionModule,
     SymbolProfileModule,
-    SymbolProfileOverwriteModule
+    SymbolProfileOverwriteModule,
+    TransformDataSourceInRequestModule
   ],
   controllers: [AdminController],
   providers: [AdminService],