+++ apps/api/src/app/symbol/symbol.module.ts
@@ -1,4 +1,5 @@
-import { ConfigurationModule } from '@ghostfolio/api/services/configuration/configuration.module';
+import { TransformDataSourceInRequestModule } from '@ghostfolio/api/interceptors/transform-data-source-in-request/transform-data-source-in-request.module';
+import { TransformDataSourceInResponseModule } from '@ghostfolio/api/interceptors/transform-data-source-in-response/transform-data-source-in-response.module';
 import { DataProviderModule } from '@ghostfolio/api/services/data-provider/data-provider.module';
 import { MarketDataModule } from '@ghostfolio/api/services/market-data/market-data.module';
 import { PrismaModule } from '@ghostfolio/api/services/prisma/prisma.module';
@@ -12,10 +13,11 @@ import { SymbolService } from './symbol.service';
   controllers: [SymbolController],
   exports: [SymbolService],
   imports: [
-    ConfigurationModule,
     DataProviderModule,
     MarketDataModule,
-    PrismaModule
+    PrismaModule,
+    TransformDataSourceInRequestModule,
+    TransformDataSourceInResponseModule
   ],
   providers: [SymbolService]
 })