+++ apps/api/src/services/data-provider/eod-historical-data/eod-historical-data.service.ts
@@ -66,7 +66,9 @@ export class EodHistoricalDataService implements DataProviderInterface {
 
   public getDataProviderInfo(): DataProviderInfo {
     return {
-      isPremium: true
+      isPremium: true,
+      name: 'EOD Historical Data',
+      url: 'https://eodhd.com'
     };
   }