+++ apps/client/src/app/pages/resources/personal-finance-tools/products.ts
@@ -66,7 +66,8 @@ export const products: Product[] = [
       'Français',
       'Italiano',
       'Nederlands',
-      'Português'
+      'Português',
+      'Türkçe'
     ],
     name: 'Ghostfolio',
     origin: $localize`Switzerland`,