+++ libs/common/src/lib/interfaces/enhanced-symbol-profile.interface.ts
@@ -1,6 +1,7 @@
 import { AssetClass, AssetSubClass, DataSource, Tag } from '@prisma/client';
 
 import { Country } from './country.interface';
+import { DataProviderInfo } from './data-provider-info.interface';
 import { ScraperConfiguration } from './scraper-configuration.interface';
 import { Sector } from './sector.interface';
 
@@ -12,6 +13,7 @@ export interface EnhancedSymbolProfile {
   countries: Country[];
   createdAt: Date;
   currency?: string;
+  dataProviderInfo?: DataProviderInfo;
   dataSource: DataSource;
   dateOfFirstActivity?: Date;
   id: string;