+++ libs/ui/src/lib/activities-table/activities-table.component.ts
@@ -78,6 +78,7 @@ export class GfActivitiesTableComponent
   @Input() dataSource: MatTableDataSource<Activity>;
   @Input() deviceType: string;
   @Input() hasPermissionToCreateActivity: boolean;
+  @Input() hasPermissionToDeleteActivity: boolean;
   @Input() hasPermissionToExportActivities: boolean;
   @Input() hasPermissionToOpenDetails = true;
   @Input() locale = getLocale();
@@ -91,10 +92,10 @@ export class GfActivitiesTableComponent
   @Input() sortDisabled = false;
   @Input() totalItems = Number.MAX_SAFE_INTEGER;
 
+  @Output() activitiesDeleted = new EventEmitter<void>();
   @Output() activityDeleted = new EventEmitter<string>();
   @Output() activityToClone = new EventEmitter<OrderWithAccount>();
   @Output() activityToUpdate = new EventEmitter<OrderWithAccount>();
-  @Output() deleteAllActivities = new EventEmitter<void>();
   @Output() export = new EventEmitter<void>();
   @Output() exportDrafts = new EventEmitter<string[]>();
   @Output() import = new EventEmitter<void>();
@@ -210,6 +211,16 @@ export class GfActivitiesTableComponent
     this.activityToClone.emit(aActivity);
   }
 
+  public onDeleteActivities() {
+    const confirmation = confirm(
+      $localize`Do you really want to delete these activities?`
+    );
+
+    if (confirmation) {
+      this.activitiesDeleted.emit();
+    }
+  }
+
   public onDeleteActivity(aId: string) {
     const confirmation = confirm(
       $localize`Do you really want to delete this activity?`
@@ -240,10 +251,6 @@ export class GfActivitiesTableComponent
     );
   }
 
-  public onDeleteAllActivities() {
-    this.deleteAllActivities.emit();
-  }
-
   public onImport() {
     this.import.emit();
   }
@@ -258,7 +265,7 @@ export class GfActivitiesTableComponent
 
   public onOpenPositionDialog({ dataSource, symbol }: UniqueAsset) {
     this.router.navigate([], {
-      queryParams: { dataSource, symbol, positionDetailDialog: true }
+      queryParams: { dataSource, symbol, holdingDetailDialog: true }
     });
   }